Improvements to let stderr output of rumor just go to the console. Python's pty.spawn directed rumor's stderr also to its stdout, which could cause rumor messages to end up in the LilyPond document.
Now a helper script is added, runpty.py, which contains some code
from Python's pty module, but leaves stderr out of the game.

This all is a workaround for installations of PyKDE that don't have the
KProcess.setUsePty method...

Also when KProcess.setUsePty can be used, we now let stderr go to
the console, instead of catching it ourselves and redirecting it to
the console anyway. (Later on, we might do something more clever with it.)

Finally, this commit adds a 'appdir' variable to lilykde.__init__.py
that points to the dir LilyKDE is installed in. This makes it easier
for other modules to find data files that belong to LilyKDE.



git-svn-id: svn://localhost/tmp/repo1/trunk@313 e61d349e-2442-45d7-aecb-3eed713e9aed